Collected the imp_travel, login_from_new_country and login_from_new_device modules into the generic detection module
