+++ buffalogs/impossible_travel/dashboard/charts.py
@@ -6,10 +6,7 @@ import pygal
 from dateutil.relativedelta import relativedelta
 from django.utils import timezone
 from impossible_travel.models import Alert, User
-from impossible_travel.modules import impossible_travel
 from pygal.style import Style
-
-imp_travel = impossible_travel.Impossible_Travel()
 
 
 def _load_data(name):
